+++ frontend/lib/identification/activation_workflow/activation_code_parser.dart
