+++ frontend/lib/identification/verification_workflow/dialogs/positive_verification_result_dialog.dart
@@ -1,7 +1,7 @@
 import 'package:ehrenamtskarte/configuration/configuration.dart';
 import 'package:ehrenamtskarte/graphql/graphql_api.dart';
 import 'package:ehrenamtskarte/identification/id_card/id_card.dart';
-import 'package:ehrenamtskarte/identification/verification/dialogs/verification_result_dialog.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/dialogs/verification_result_dialog.dart';
 import 'package:ehrenamtskarte/proto/card.pb.dart';
 import 'package:flutter/material.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
