+++ frontend/lib/identification/verification_workflow/dialogs/verification_result_dialog.dart
