+++ frontend/lib/identification/verification_workflow/query_server_verification.dart
@@ -1,5 +1,5 @@
 import 'package:ehrenamtskarte/graphql/graphql_api.dart';
-import 'package:ehrenamtskarte/identification/verification/verification_hasher.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/verification_hasher.dart';
 import 'package:ehrenamtskarte/proto/card.pb.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
