+++ frontend/lib/identification/verification_workflow/verification_hasher.dart
