+++ frontend/lib/identification/verification_workflow/verification_qr_code_processor.dart
@@ -1,5 +1,5 @@
 import 'package:ehrenamtskarte/identification/qr_code_scanner/qr_code_processor.dart';
-import 'package:ehrenamtskarte/identification/verification/scanner/verification_qr_content_parser.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/verification_qr_content_parser.dart';
 import 'package:ehrenamtskarte/proto/card.pb.dart';
 
 DynamicVerifyCode processQrCodeContent(String rawBase64Content) {
