+++ frontend/lib/identification/verification_workflow/verification_qr_content_parser.dart
