+++ frontend/lib/identification/verification_workflow/verification_qr_scanner_page.dart
@@ -6,12 +6,12 @@ import 'package:ehrenamtskarte/identification/activation_code_model.dart';
 import 'package:ehrenamtskarte/identification/otp_generator.dart';
 import 'package:ehrenamtskarte/identification/qr_code_scanner/qr_code_processor.dart';
 import 'package:ehrenamtskarte/identification/qr_code_scanner/qr_code_scanner_page.dart';
-import 'package:ehrenamtskarte/verification/dialogs/internet_connection_verification_dialog.dart';
-import 'package:ehrenamtskarte/identification/verification/dialogs/negative_verification_result_dialog.dart';
-import 'package:ehrenamtskarte/identification/verification/dialogs/positive_verification_result_dialog.dart';
-import 'package:ehrenamtskarte/identification/verification/dialogs/verification_info_dialog.dart';
-import 'package:ehrenamtskarte/identification/verification/query_server_verification.dart';
-import 'package:ehrenamtskarte/identification/verification/verification_qr_code_processor.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/dialogs/internet_connection_verification_dialog.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/dialogs/negative_verification_result_dialog.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/dialogs/positive_verification_result_dialog.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/dialogs/verification_info_dialog.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/query_server_verification.dart';
+import 'package:ehrenamtskarte/identification/verification_workflow/verification_qr_code_processor.dart';
 import 'package:ehrenamtskarte/proto/card.pb.dart';
 import 'package:ehrenamtskarte/widgets/navigation_bars.dart' as nav_bars;
 import 'package:flutter/material.dart';
